[AND-867] Gate Play & Earn on locally detected country
Firebase geo-targeting of show_play_and_earn is not trusted on its own:
shouldShowPlayAndEarn() now also requires the locally detected country
(network -> SIM -> locale, via the new DeviceCountryProvider) to be in
the allowed list. The list comes from the pae_countries remote config
value (JSON array of ISO codes), falling back to a hardcoded set when
missing or unparseable; unknown country hides the feature.
